Initial push of C++ files.
The code in this project uses QT. The project was built in QT Creator, where the GUI was designed by moving components around by eye rather than in code.
This code loads images, some of which contain Waldo and some that do not, then applies various filters in order to extract feature information from each image. This project uses decision trees to classify an image as containing or not containing Waldo. Thus, it is imperative to convert the original feature space--being 64 x 64 x 3 (pixel dimensions times rgb color space)--to a set of features that can better be used as binary decisions as nodes of the tree. Three features are extracted based on the filters applied: whether the image contains red and white stripes, whether it contains black glasses, and whether it contains blue pants. These features are used to dynamically construct a decision tree (with no pruning since the max depth is shallow) on a subset of the images, which is the training set. Then the testing set (remaining